requirements.txt
python-dotenv
flask
selenium
webdriver-manager
pytest
flask-dance[sqla]
Flask-SQLAlchemy
flask-login
Flask-Assets
cssmin
blinker
applicationinsights